requirements.txt
Django==1.11
psycopg2==2.7.3.2
django-extensions==1.9.9
django-autoslug==1.9.3
Pillow==5.0.0
django-password-reset==1.0
stripe==1.77.1
django-solo==1.1.3